requirements.txt
APScheduler==3.6.3 #                 python-telegram-bot
attrs==22.2.0 #                      pytest
cachetools==4.2.2 #                  python-telegram-bot
certifi==2022.12.7 #                 python-telegram-bot, requests
charset-normalizer==2.1.1 #          requests
exceptiongroup==1.1.0 #              pytest
greenlet==2.0.1 #                    sqlalchemy
idna==3.4 #                          requests
ijson==3.1.4 #                       ijson
iniconfig==1.1.1 #                   pytest
packaging==22.0 #                    pytest
pluggy==1.0.0 #                      pytest
pydantic==1.10.4 #                   pydantic
pytest==7.2.0 #                      pytest
python-dotenv==0.21.0 #              python-dotenv
python-telegram-bot==13.15 #         python-telegram-bot
pytz==2022.7 #                       python-telegram-bot
pytz-deprecation-shim==0.1.0.post0 # python-telegram-bot
requests==2.28.1 #                   requests
six==1.16.0 #                        python-telegram-bot
sqlalchemy==1.4.46 #                 sqlalchemy
sqlalchemy-stubs==0.4 #              sqlalchemy-stubs
tomli==2.0.1 #                       pytest
tornado==6.1 #                       python-telegram-bot
typing_extensions==4.4.0 #           pydantic
tzdata==2022.7 #                     python-telegram-bot
tzlocal==4.2 #                       python-telegram-bot
urllib3==1.26.13 #                   requests
